# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Video Game Rental API adalah sistem backend untuk platform penyewaan game fisik",
        "title": "Video Game Rental API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of all payments (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Payments"
                ],
                "summary": "Get all payments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payments retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/payments/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of payments filtered by status (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Payments"
                ],
                "summary": "Get payments by status",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "completed",
                            "failed",
                            "refunded"
                        ],
                        "type": "string",
                        "description": "Payment status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payments retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/bookings/{booking_id}/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get payment information for a specific booking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment by booking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid booking ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create payment for a booking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment created successfully",
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/payments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed payment information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid payment ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/webhooks/payments": {
            "post": {
                "description": "Receive payment status updates from payment provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Payment webhook",
                "parameters": [
                    {
                        "description": "Webhook payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PaymentWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid webhook payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "provider"
            ],
            "properties": {
                "provider": {
                    "enum": [
                        "stripe",
                        "midtrans"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.PaymentProvider"
                        }
                    ]
                }
            }
        },
...
        "dto.PaymentWebhookRequest": {
            "type": "object",
            "required": [
                "provider_payment_id",
                "status"
            ],
            "properties": {
                "failure_reason": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "provider_payment_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.ApprovalStatus": {
            "type": "string",
            "enum": [
                "pending_approval",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "ApprovalPending",
                "ApprovalApproved",
                "ApprovalRejected"
            ]
        },
        "model.Booking": {
            "type": "object",
            "required": [
                "end_date",
                "start_date"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_price": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "game": {
                    "$ref": "#/definitions/model.Game"
                },
                "game_id": {
                    "type": "integer"
                },
                "handover_confirmed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "partner": {
                    "$ref": "#/definitions/model.User"
                },
                "partner_id": {
                    "type": "integer"
                },
                "payment": {
                    "$ref": "#/definitions/model.Payment"
                },
                "rental_days": {
                    "type": "integer"
                },
                "return_confirmed_at": {
                    "type": "string"
                },
                "review": {
                    "$ref": "#/definitions/model.Review"
                },
                "security_deposit": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BookingStatus"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_rental_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.BookingStatus": {
            "type": "string",
            "enum": [
                "pending_payment",
                "confirmed",
                "active",
                "completed",
                "cancelled",
                "disputed"
            ],
            "x-enum-varnames": [
                "BookingPendingPayment",
                "BookingConfirmed",
                "BookingActive",
                "BookingCompleted",
                "BookingCancelled",
                "BookingDisputed"
            ]
        },
        "model.Category": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.Game": {
            "type": "object",
            "required": [
                "name",
                "rental_price_per_day"
            ],
            "properties": {
                "approval_status": {
                    "$ref": "#/definitions/model.ApprovalStatus"
                },
                "approved_at": {
                    "type": "string"
                },
                "approved_by": {
                    "type": "integer"
                },
                "approver": {
                    "$ref": "#/definitions/model.User"
                },
                "available_stock": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "category_id": {
                    "type": "integer"
                },
                "condition": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "partner": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.User"
                        }
                    ]
                },
                "partner_id": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "rental_price_per_day": {
                    "type": "number"
                },
                "security_deposit": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 1
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "booking": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Booking"
                        }
                    ]
                },
                "booking_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "failed_at": {
                    "type": "string"
                },
                "failure_reason": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/model.PaymentProvider"
                },
                "provider_payment_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.PaymentStatus"
                }
            }
        },
        "model.PaymentProvider": {
            "type": "string",
            "enum": [
                "stripe",
                "midtrans"
            ],
            "x-enum-varnames": [
                "ProviderStripe",
                "ProviderMidtrans"
            ]
        },
        "model.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "paid",
                "failed",
                "refunded"
            ],
            "x-enum-varnames": [
                "PaymentPending",
                "PaymentPaid",
                "PaymentFailed",
                "PaymentRefunded"
            ]
        },
        "model.Review": {
            "type": "object",
            "required": [
                "rating"
            ],
            "properties": {
                "booking": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Booking"
                        }
                    ]
                },
                "booking_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "game": {
                    "$ref": "#/definitions/model.Game"
                },
                "game_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "required": [
                "email",
                "full_name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/model.UserRole"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.UserRole": {
            "type": "string",
            "enum": [
                "customer",
                "partner",
                "admin",
                "super_admin"
            ],
            "x-enum-varnames": [
                "RoleCustomer",
                "RolePartner",
                "RoleAdmin",
                "RoleSuperAdmin"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}